+++ elphick/mass_composition/stream.py
@@ -1,4 +1,4 @@
-from typing import Optional, Callable
+from typing import Optional, Callable, Generator
 
 from elphick.mass_composition import MassComposition
 
@@ -47,17 +47,18 @@ class Stream(MassComposition):
         mc1, mc2 = super().split(fraction, name_1, name_2)
         return Stream.from_mass_composition(mc1), Stream.from_mass_composition(mc2)
 
-    def apply_partition(self, definition: Callable,
-                        name_1: Optional[str] = None, name_2: Optional[str] = None) -> tuple['Stream', 'Stream']:
+    def split_by_partition(self, partition_definition: Callable,
+                           name_1: Optional[str] = None, name_2: Optional[str] = None) -> Generator[
+        'Stream', None, None]:
         """
         Partition the object along a given dimension.
 
         This method applies the defined separation resulting in two new objects.
 
-        See also: split
+        See also: split, split_by_function
 
         Args:
-            definition: A partition function that defines the efficiency of separation along a dimension
+            partition_definition: A partition function that defines the efficiency of separation along a dimension
             name_1: The name of the reference stream created by the split
             name_2: The name of the complement stream created by the split
 
@@ -66,8 +67,56 @@ class Stream(MassComposition):
 
 
         """
-        mcs = super().apply_partition(definition, name_1, name_2)
-        return (Stream.from_mass_composition(mc) for mc in mcs)
+        streams = super().split_by_partition(partition_definition, name_1, name_2)
+        return (Stream.from_mass_composition(stream) for stream in streams)
+
+    def split_by_function(self, split_function: Callable,
+                          name_1: Optional[str] = None,
+                          name_2: Optional[str] = None) -> Generator['Stream', None, None]:
+        """Split an object using a function.
+
+        This method applies the function to self, resulting in two new objects. The object returned with name_1
+        is the result of the function.  The object returned with name_2 is the complement.
+
+        See also: split, split_by_estimator, split_by_partition
+
+        Args:
+            split_function: Any function that transforms the dataframe from a MassComposition object into a new
+             dataframe with values representing a new (output) stream.  The returned dataframe structure must be
+             identical to the input dataframe.
+            name_1: The name of the stream created by the function
+            name_2: The name of the complement stream created by the split, which is calculated automatically.
+
+        Returns:
+            A generator of two Streams,
+
+
+        """
+        streams = super().split_by_function(split_function, name_1, name_2)
+        return (Stream.from_mass_composition(stream) for stream in streams)
+
+    def split_by_estimator(self, estimator: 'sklearn.base.BaseEstimator',
+                           name_1: Optional[str] = None,
+                           name_2: Optional[str] = None) -> Generator['Stream', None, None]:
+        """Split an object using a sklearn estimator.
+
+        This method applies the function to self, resulting in two new objects. The object returned with name_1
+        is the result of the estimator.predict() method.  The object returned with name_2 is the complement.
+
+        See also: split, split_by_function, split_by_partition
+
+        Args:
+            estimator: Any sklearn estimator that transforms the dataframe from a MassComposition object into a new
+             dataframe with values representing a new (output) stream using the predict method.  The returned
+             dataframe structure must be identical to the input dataframe.
+            name_1: The name of the stream created by the estimator.
+            name_2: The name of the complement stream created by the split, which is calculated automatically.
+
+        Returns:
+            tuple of two datasets, the first with the mass fraction specified, the other the complement
+        """
+        streams = super().split_by_estimator(estimator, name_1, name_2)
+        return (Stream.from_mass_composition(stream) for stream in streams)
 
     def add(self, other: 'Stream', name: Optional[str] = None) -> 'Stream':
         """